+++ packages/dock-blockchain-modules/tests/integration/anoncreds/accumulator.test.js
@@ -26,10 +26,7 @@ import {
 import { DockDid, DockAccumulatorId } from "@docknetwork/credential-sdk/types";
 
 import { AccumulatorType } from "@docknetwork/credential-sdk/modules/abstract/accumulator";
-import {
-  AbstractAccumulatorModule,
-  MultiApiAccumulatorModule,
-} from "@docknetwork/credential-sdk/modules";
+import { AbstractAccumulatorModule } from "@docknetwork/credential-sdk/modules";
 import { getAllEventsFromBlock } from "@docknetwork/dock-blockchain-api/utils/chain-ops";
 import { registerNewDIDUsingPair } from "../helpers";
 import {
@@ -484,7 +481,7 @@ describe("Accumulator Module", () => {
         chainModule.dockOnly.constructor.parseEventAsAccumulatorUpdate(
           events1[0].event
         )
-      ).toEqual([id[1].value, accumulated2]);
+      ).toEqual([id.asDock.value, accumulated2]);
 
       const updates2 = await chainModule.dockOnly.getUpdatesFromBlock(
         id,
@@ -506,7 +503,7 @@ describe("Accumulator Module", () => {
         chainModule.dockOnly.constructor.parseEventAsAccumulatorUpdate(
           events2[0].event
         )
-      ).toEqual([id[1].value, accumulated3]);
+      ).toEqual([id.asDock.value, accumulated3]);
 
       const updates3 = await chainModule.dockOnly.getUpdatesFromBlock(
         id,