+++ examples/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @docknetwork/sdk-examples
 
+## 0.3.0
+
+### Minor Changes
+
+- Implement `MultiApi` modules allowing to use different APIs in the same moment
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/dock-blockchain-modules@0.4.0
+  - @docknetwork/dock-blockchain-api@0.3.0
+  - @docknetwork/credential-sdk@0.4.0
+
 ## 0.2.0
 
 ### Minor Changes